Demodulize annotation policies
