+++ spec/factories/annotation_factory.rb
@@ -1,17 +1,17 @@
 # frozen_string_literal: true
 
 FactoryGirl.define do
-  factory :annotation, :class => Annotations::Annotation do
+  factory :annotation, :class => Annotation do
     content_item_id { Faker::Number.number 2 }
     user { build :user }
     deck { build :deck }
 
-    factory :conversation, :class => Annotations::Conversation do
+    factory :conversation, :class => Conversation do
       comment_type { %i[question note].sample }
       text { Faker::Lorem.sentence(3) }
     end
 
-    factory :comment, :class => Annotations::Comment do
+    factory :comment, :class => Comment do
       text { Faker::Lorem.sentence(3) }
 
       conversation do
